fix(sessions): repair golangci-lint findings in turns/session paths
Resolve every golangci-lint v2.12.2 finding in the owned HA-63/HA-64
session paths without changing product semantics:

- materializer Run: only treat a pass abort whose error actually
  unwraps to the run context's own cancellation (errors.Is(err,
  ctx.Err())) as a clean nil exit; a genuine store/source failure is
  reported even when cancellation lands concurrently (nilerr real
  error path, no suppression).
- updateTurn: drop the never-used TurnRow result (every caller
  discarded it); rootTurn/taskTurn/rootTaskTurn drop the never-used
  TurnID result; applyPauseResumed drops the unused payload param.
- projector.go: wrap the underlying validateText error with %w
  (Go 1.20 multi-%w) instead of %v — message text unchanged,
  ErrInvalidInput stays errors.Is-reachable and the cause is now
  reachable too.
- conformancetest: check LoadCheckpoint/SaveCheckpoint errors instead
  of blank-discarding; derive concurrent checkpoint values in the
  uint64 domain (no int->uint64 conversion, gosec G115); simplify the
  bool comparison.
- Tests: mechanical intrange/copyloopvar/govet-inline/unconvert
  conversions, unparam parameter removals where every caller passed
  the same value, and removal of genuinely unused helpers
  (toolFailedEv, mixedCatalog, fakeClock.advance, testStore.saveRow).

Lifecycle counters, turn paging/snapshot semantics, durable erasure,
materializer monotonicity, safe observation data, and Protocol
authorization are unchanged.

+++ internal/sessions/turns/materializer/usage_arithmetic_test.go
@@ -53,7 +53,7 @@ func TestMaterialize_UsageArithmeticFailsLoud(t *testing.T) {
 		{"-Inf cost", nil, map[string]any{"TotalCost": math.Inf(-1)}},
 		{"out-of-range token float", map[string]any{"PromptTokens": 1e300}, nil},
 		{"non-numeric cost", nil, map[string]any{"TotalCost": "twelve"}},
-		{"latency ms to ns overflow", map[string]any{"LatencyMS": int64(math.MaxInt64/int64(time.Millisecond)) + 1}, nil},
+		{"latency ms to ns overflow", map[string]any{"LatencyMS": int64(math.MaxInt64/time.Millisecond) + 1}, nil},
 	}
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
